+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Parse model position
+- Parse model name
 - Better validation for Unity meta files in build pipeline
 
 ## [2.1.2] - 2023-01-03